Hide unmatched branches during search
As search continues, hide sections of the tree with no matching children.
e.g. staring to type "top" might match several parts of the tree, so they are opened.
Then the user continues to type "topology". This no longer matches some nodes, but currently those branches remain open. In a large tree, this can be distracting.
(This behaviour could be an option defined in _options)